% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_KDE.R
\name{calculate_KDE}
\alias{calculate_KDE}
\title{Get KDE distribution}
\usage{
calculate_KDE(data_input, band = 1, grid_size = 1000)
}
\arguments{
\item{data_input}{numeric vector of measurement values}

\item{band}{kernel bandwidth smooth param. higher is smoother}

\item{grid_size}{the number of equally spaced points at which to estimate the density.}
}
\value{
dataframe of x and y coordinates
}
\description{
Get KDE distribution
}
