% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_quantogram.R
\name{calculate_modular_quantogram}
\alias{calculate_modular_quantogram}
\title{calculate Kendall's modular quantogram (cosine quantogram but accounting for the possibility of an offset}
\usage{
calculate_modular_quantogram(x, mqparams = getOption("CONSTANTS_QUANTOGRAM"))
}
\arguments{
\item{x}{numeric vector of measurement values}

\item{params}{cosine quantogram parameters including "RNG_START", "RNG_END", "STEP", "Q_MIN"}
}
\value{
two columned dataframe of quanta (q) and their scores (f_q)
}
\description{
calculate Kendall's modular quantogram (cosine quantogram but accounting for the possibility of an offset
}
