% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_quantogram.R
\name{calculate_sine_quantogram}
\alias{calculate_sine_quantogram}
\title{the cosine quantogram but with sines, used to calculate Kendall's modular quantogram}
\usage{
calculate_sine_quantogram(x, params = getOption("CONSTANTS_QUANTOGRAM"))
}
\arguments{
\item{x}{numeric vector of measurement values}

\item{params}{cosine quantogram parameters including "RNG_START", "RNG_END", "STEP", "Q_MIN"}
}
\value{
two columned dataframe of quanta (q) and their scores (f_q)
}
\description{
the cosine quantogram but with sines, used to calculate Kendall's modular quantogram
}
