% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo_test.R
\name{monte_carlo_test_modular}
\alias{monte_carlo_test_modular}
\title{sample the (non quantal) KDE repeatedly and take modular quantograms of the samples, in order to determine the significance of quantogram scores}
\usage{
monte_carlo_test_modular(
  data_input,
  kde,
  num_simulations,
  peakHeight = 0,
  significance = 0.95
)
}
\arguments{
\item{data_input}{numeric vector of measurement values (unrounded)}

\item{kde}{the kde generated for the data (use calculate KDE)}

\item{num_simulations}{how many Monte Carlo simulations are run}

\item{peakHeight}{specific quantogram peak height to find the significance of (optional)}

\item{significance}{specific significance to find the minimum peak height required for (optional)}
}
\value{
a list of 4 containing: dataframe of all max peak heights in simulations, table of significances, and the values values associated with peakHeight and significance respectively
}
\description{
sample the (non quantal) KDE repeatedly and take modular quantograms of the samples, in order to determine the significance of quantogram scores
}
