% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bootstrap.R
\name{plot_bootstrap}
\alias{plot_bootstrap}
\title{plot bootstrap produced by compute_bootstrap (if something else is passed in, it will just print() whatever it is)}
\usage{
plot_bootstrap(bootstrap)
}
\arguments{
\item{bootstrap}{expecting a list from compute_bootstrap}
}
\value{
nothing
}
\description{
plot bootstrap produced by compute_bootstrap (if something else is passed in, it will just print() whatever it is)
}
