% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calculate_selective_least_squares.R
\name{plot_calculate_selective_least_squares}
\alias{plot_calculate_selective_least_squares}
\title{Plot and calculate SLS results (for whole range of possible quanta)}
\usage{
plot_calculate_selective_least_squares(
  x,
  sector_name,
  params = getOption("CONSTANTS_SLS")
)
}
\arguments{
\item{x}{numeric vector of measurement values}

\item{sector_name}{label added to plot}

\item{params}{(optional) default is set to use CONSTANTS_SLS option}
}
\description{
Plot and calculate SLS results (for whole range of possible quanta)
}
