% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quantogram.R
\name{plot_quantogram}
\alias{plot_quantogram}
\title{Plot cosine quantogram using ggplot}
\usage{
plot_quantogram(df_results, sector_label)
}
\arguments{
\item{df_results}{this can be the quantogram generated by calculate_quantogram}

\item{sector_label}{label on the graph}
}
\value{
ggplot object
}
\description{
Plot cosine quantogram using ggplot
}
