% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_bootstrap.R
\name{sample_bootstrap}
\alias{sample_bootstrap}
\title{Get bootstrap sample}
\usage{
sample_bootstrap(x, decrease_by = 0)
}
\arguments{
\item{x}{numeric vector of measurement values}

\item{decrease_by}{size of sample = length of x - decrease_by}
}
\value{
the numeric vector of samples
}
\description{
Get bootstrap sample
}
\examples{
SW1 = c(10.8, 11.2, 12.7, 16.4, 16.9, 20.5) SWsample = sample_bootstrap(SW1, decrease_by = 2)
}
