% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unround_data.R
\name{unround_data}
\alias{unround_data}
\title{Unrounding method}
\usage{
unround_data(input, depth = 0.01)
}
\arguments{
\item{input}{a vector of measurements}

\item{depth}{the accuracy, analogous to smallest significant digit}
}
\value{
the vector input, unrounded (random number b/w -1/2 depth and 1/2 depth)
}
\description{
Unrounding method
}
\examples{
data = c(1,2,3) unroundedData = unround_data(data)
}
