% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_constants.R
\name{update_constants}
\alias{update_constants}
\title{update the constants used when calculating the cosine quantogram (and some other functions)}
\usage{
update_constants(rngStart, rngEnd, stepSize, Qmin)
}
\arguments{
\item{rngStart}{where the quantum search starts}

\item{rngEnd}{where the search ends}

\item{stepSize}{the increment size of the calculation}

\item{Qmin}{I have no clue what this is}
}
\description{
update the constants used when calculating the cosine quantogram (and some other functions)
}
