% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_constants.R
\name{update_constants_sls}
\alias{update_constants_sls}
\title{update the constants used when calculating the SLS}
\usage{
update_constants_sls(slsrngStart, slsrngEnd, stepSize, slsMaxNumQuanta)
}
\arguments{
\item{slsrngStart}{where the sls search starts}

\item{slsrngEnd}{where the sls search ends}

\item{stepSize}{the increment size of the calculation}

\item{slsMaxNumQuanta}{how many quanta should be looked for max}
}
\description{
update the constants used when calculating the SLS
}
