% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unround_data.R
\name{unround_data}
\alias{unround_data}
\title{Unrounding method}
\usage{
unround_data(input, depth = 0.01)
}
\arguments{
\item{input}{numeric vector of measurement values}

\item{depth}{the accuracy, analogous to smallest significant digit}
}
\value{
the unrounded numeric vector (random number b/w -1/2 depth and 1/2 depth)
}
\description{
Unrounding method
}
